improve test debugging logs
